1,增加进入对战的匹配界面 2，增加对战的准备设置 3，修正达到上限后不能增加出战武将

+++ Classes/Dlg/Main/PnlHero.cpp
@@ -89,7 +89,7 @@ void PnlHero::updateUI()
 void PnlHero::updateSvHero() {
 	map<int, ValueMap >& myObj = *(DBM()->getMyObj());
 	int i = 0;
-	int rowCount = ceil(myObj.size() / 3);
+	int rowCount = ceil((float)myObj.size() / 3);
 	Size svInnerSize(574, (254 + 6) * rowCount + 6);
 	this->sv_hero->setInnerContainerSize(svInnerSize);
 	svInnerSize = this->sv_hero->getInnerContainerSize();
